refactor(editor): Extract orchestration code into features (no-changelog)
