+++ packages/frontend/editor-ui/src/features/orchestration.ee/components/WorkerAccordion.vue
