+++ packages/frontend/editor-ui/src/features/orchestration.ee/components/WorkerJobAccordion.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { RunningJobSummary } from '@n8n/api-types';
-import WorkerAccordion from './WorkerAccordion.ee.vue';
+import WorkerAccordion from './WorkerAccordion.vue';
 import { useI18n } from '@n8n/i18n';
 
 import { N8nText } from '@n8n/design-system';
